Initial commit. Working source build.
Framework for future development. A few scripts to build a base
container and run a source build (as opposed to a Debian package build).

Configured in vars.sh. This will be used for customisations.

The base container installs packages, builds a newer ccache, and runs
install-deps.sh out of the git source directory. This means it will
adapt to your checked-out version. We'll need to change this slightly so
that having two runs with different Ceph versions doesn't invalidate the
image cache.