Support multiple images based on the source.
One of the advantages of the base image is that `./install-deps.sh` has
been run already. This is somewhat less useful if the install-deps is
wrong; and this will be true if people are building multiple versions of
Ceph, which *is* going to happen.

Hack around it. We build the preinstall/ directory at build time from
the git working directory. It needs to be in this directory for `docker
build` to work, so wrap the manipulation of that directory in a flock(1)
to prevent multiple builds out of this directory from stepping on each
other. Take a sha256 checksum of the preinstall/ directory and use (in
truncated form) that hash as the image tag.

When running a build, we recreate the same preinstall/ directory (this
time in a temporary dir) and generate the same hash. This matches our
base build image exactly with our source.