Custom dependency infrastructure.
In build/ there are scripts and some common config for building custom
dependencies and tools. Now this includes ccache, abseil-cpp, grpc and
openssl v3, though only ccache is currently enabled. (The flags to enable
them are in build/config.env .)

Once gRPC is integrated, it will be built once for the container, and
then linked as radosgw is built. If the version needs to change, it's in
the Docker setup.

Same for OpenSSL v3. We don't even know if this is going to work, but
the infrastructure exists to make it easy to try.