Install golang and ccache from tarballs.
The ccache source install is there but there's really no need for it, if we
assume amd64 anyway. (Sorry no Docker for Mac support.)

Both are straightforward binary installs. The ccache installation pulls down a
specific gpg key instead of automatically retrieving the key, since trusting
the .asc file to give the right key is kinda dangerous. (I think it's a fringe
concern.)

Golang is the newest non-zero stable release. It doesn't really matter to us
much, but we need go to install buf, which is used to generate protobuf
sources. Using the standard gRPC tools to manage protobufs is nothing but
pain, so we're not doing it.