Don't need to both copy and mount /tools.
Pick one. It's convenient to be able to edit the tools in place when
working on the builder, but most of the time it makes more sense to just
copy them in, since we're already copying loads of stuff.

It'll get re-copied if it changes anyway, it's just an extra step when
working on the tools themselves.