Big install-deps.sh versioning update.
We now version the image tag on both the preinstall environment and the git
short ref of the akceph-build repo. This is a future-proofing check against
changes to the build tool not being reflected in the containers people run.

This means we'll get rebuilds every time something is committed to the
akceph-build repo. That's a nuisance for me, but for most people I think it
won't matter. It's safer to have rebuilds and keep the images up to date.

The previous changes worked, but would cause install-deps.sh rebuilds every
time. It turns out you can't use a tmp dir for files copied in, because you
have to pass the tmp dir name, that's in an ARG, and if the ARG is different,
the cache is invalidated for any subsequent Dockerfile commands that refer to
the ARG. This is sensible, but we have to work around it.

I'm doing that by having the container builder create the preinstall in a
fixed directory containing the preinstall's hash. There's a chicken-and-egg
here, because we need the preinstall directory to know the hash. The inelegant
but simple solution is to just build the preinstall twice. Of course we could
move it instead, but it's so quick to create that I don't think it's worth the
trouble.