Infer the release folder name, with some checks.
It's a nuisance to have to specify -r FOLDER to set the release folder name
when using -s REF, because the folder will usually be release_REF. So by
default we take ref, transform it if necessary to be a valid folder name, and
set that as the release folder name.

You can still override it if you like, e.g. if CI expects a particular folder,
but in most cases the default will be useful.