Revert "Attempt to patch out DWZ compression."
This reverts commit ea1178d9056ea0270a4678ec8447e90159932dae.

This can't work - the Debian builds use (see ceph
bin/git-archive-all.sh) `git archive` to create the tarballs, which will
ignore source edits and build, well, an archive of the current committed
repo. Patching the working copy doesn't do anything.

If we want to speed up this build by disabling dwz, it's a two-line
patch but it will have to be a patch.