Attempt to detect a detached-head working copy.
If the branch looks like 'HEAD', chances are we're on a detached HEAD.
There's a way to look up the branch from there, let's try it.

Note this won't fix the unhelpful version given to the packages
themselves - this is dictated by the Ceph build scripts and is based on
`git describe` output. It will just affect the rpmbuild_ paths.