Set RPM build macros. Disable FBP.
Create a ~/.rpmmacros when building the RPMs. Set the %packager, which
will show allow people to see where the packages come from.

More importantly, do `%define _unpackaged_files_terminate_build 0` to
disable the 'feature' that anything we didn't explicitly package will
break the build. This is referred to as the 'Fascist Build Policy' (e.g.
https://jfearn.fedorapeople.org/en-US/RPM/4/html/RPM_Guide/ch09s05s07.html)
and I understand what it's for, I understand why it's called back, and I
know I don't really care that much about disabling it.

Do I think upstream Ceph should disable it? No - it makes sense there.
For us it does make some sense - at least for now. The alternative would
be packaging Expat, which feels like overkill since it's linked into the
binaries that are already packaged.

It might cause problems when debugging. If so, we'll revisit this.