Attempt to kludge the web server IP address.
The lab box doesn't have an entry in the DNS, its hostname only appears
to exist in /etc/hosts. This is fine on the host itself, but a container
will rely on the DNS or 'some other means'. Typically one would solve
this with orchestration such as k8s or docker compose. Try to hack it
for now.

So: Resolve the IP address using getent, filter for IPv4, and use what
we find. This will break at some point, e.g. I'm sure github actions
won't care for it.