Add an image with the debug symbols preinstalled.
It's a pain to get the debug symbols installed in the CentOS Stream
image. Create an image at build time that already has them installed.

This is a quite small adjustment, as the packages are already present.
There will now be an additional docker image, akdaemon-gen2-debug, which
is the same as akdaemon-gen2 except that all the debuginfo RPMs are
installed. This is *much* easier to do at build time, as we've not yet
surrendered root in the container and the temporary package web server
is still running. It's really just one additional 'yum install' command
and some plumbing.

There's also a developer option -P to official-container.sh, which has
reasonable help notes. It's to be used when working on akceph-build, it
is unlikely to be useful outside that context.