Add a way to not delete the RPM build container.
Normally (and correctly) we'd use --rm in the 'docker run' command line
so a container is deleted after use. Otherwise the disk use grows
without bound.

However, when debugging a build system, it's very annoying to not have
any way to see a partial build if it fails. Using the new -l option
means a container is left around after running, and can be inspected.